Added assess_status() functionality - no tests yet
This adds the assess_status functionality, but without any tests yet.
It also adds a register function to enable the release to be specified
before an OpenStackCharm class is instantiated, thus providing the
mechanism to choose the class to use for a release (or series of
releases).  This was missing from earlier commits.