API specific RelationAdapters class  & restart fix
API service specifc code had landed in OpenStackRelationAdapters.
This change adds a child class of OpenStackRelationAdapters called
OpenStackAPIRelationAdapters and moved api specific code into it.

All services need to be stopped and then all services need to be
restarted rather than restarting each service at a time. This is
to deal with the situation where ports are being changed and a
running service maybe bound to a port a new service is intending
too use.

Change-Id: I57991b9d5be82c6d72cba6f53a27172bd869795c