Add 'declarative helpers'
Declarative helpers takes (most) of the boiler plate of writing an
OpenStack charm and places it into this module.  This simplified the
writing, and testing, of an OpenStack charm when using reactive and the
layered approach.

This change makes a BREAKING change to existing charms written using
charms.openstack.  This is that the 'assess-status' hook that was
originally in layer-openstack has been removed and made optional via the
declarative helpers.  This was to maintain consistency with the
declarative helpers and also to allow charm authors to write their own
'assess-status' state handler as required.

Thus there is also a change to (charm-)layer-openstack which removes the
wired in handler under the same topic.

Change-Id: I3c74f60bb4ed7901828902118697f310622c4061