Add application version support
Juju 2.0 allows charms to specify the version of the application
they are deploying, which is provided to users via the Juju status
output.

Add attribute to base charm class and update assess_status method
to base OpenStackCharm class; version is detected using the following
process:

 1) charm.version_package (if set for backwards compat)
 2) charm.packages[0]
 3) os_release() (typically codename for OS release)

This should automatically enable application version in reactive
OpenStack charms once rebuilt with new charms.openstack.

NOTE: the get_upstream_version helper is present in the trunk
branch of charmhelpers, but is not yet released; this change
uses a local copy of this function to workaround this timing
issue, which can be dropped as soon as 0.9.2 of ch releases.

Change-Id: I00029e26462e8196daa1fe157831b649bdfe7113