Add 'options' to OpenStackCharm class
This changeset adds an 'options' property to the OpenStackCharm
class that is evaluated/added when the class is created.  This
enables the OpenStackCharm derived class to access computed
options as well as the config from the config.yaml.

Addtional changes are to simplify testing (and the need to mock)
by delaying instantiating or evaluating properties/members on
the the class during the __init__() method.  This means that
they are lazily evaluated which also helps with hook performance
if the property is not used during that hook invocation.

Change-Id: If6d103b4f62c95b0fa76562a18e418e0d319e987