charms.openstack charms don't open ports
This patchset adds a new method to the OpenStackCharm() class called
'update_api_ports()' which, by default, uses the api_ports static member
of the same class to expose the ports using the Juju 'open_port' and
'close-port' commands. This method will open any ports that aren't
already open and close any ports that shouldn't be open according to the
Juju 'opened-ports' command.

The 'api_ports' static member can be overriden by providing a simple
list (or other iterable) to the method call.

By default, the default 'install' method on the OpenStackCharm() class
now calls update_api_ports() after it has installed the packages.  This
is normally what a charm would want to do: install the software, which
usually results in it starting, and then expose its ports. In order to
disable the default behaviour, it is necessary to completely override
the default install() method.  This is intentional, as the charm should
normally expose its api_ports.

Also, the update_api_ports() is called from the (new) update_charm()
method which can be accessed using a defaults 'update-charm'.  This is
so that ports can be changed when a charm is upgraded.

Change-Id: I4c61179cc6217f4407340a9205619c04b35661a9
Closes-Bug: 1646089