Enable Memcache for API svc token caching
On deployments of API charms using OpenStack release Mitaka or later
enable memcache for token caching.

With the release of 4.2.0 of keystonemiddleware using the
in-process token cache is no longer recommended. It is recommended
that a memcache backend to store tokens is used instead,

This installs and configures memcache for API charms
to use memcache for token caching.

http://docs.openstack.org/releasenotes/keystonemiddleware/mitaka.html#id2

Create os_release_data module for storing config like
KNOWN_RELEASES as this is not tied to the charm module an is useful
elsewhere

Change-Id: I6790f7968d0bd0493f103aec4b0b2a46333af6b0