Use charmhelpers OPENSTACK_RELEASES instead of internal KNOWN_RELEASES
As part of the fix-alpha-comparisons-1659575 topic, OpenStack reactive
charms that use charms.openstack don't generally need an update.
However, it was noticed that charms.openstack used its own list, as it
predated the charm-helpers version.  In order to only have to maintain
the list in one place (contrib.openstack.utils.OPENSTACK_RELEASES) this
changeset removes the internal list and uses the external list.

The py27 test has been disabled as the charms.reactive library has
become py3.5+ only.  However, the gate retains the test (policy) so we
need to just disable the py27 here in the tox.ini.

Change-Id: Iaa08252e5093f69a201af2a61cbe952d3ab3b171
Related-Bug: #1659575