Remove py2.7 support
The underlying library (charms.reactive) no longer support py27, and
reactive charms are only py35 and later supported.  Thus, there is no
reason to support py27 in charms.openstack.

Note: the zuul gate still has a py27 test so that is left in the tox.ini
but it simply evaluates to true.

Change-Id: I9a4e18038b706286f60e64fcf67f3e63a90feda9