Ensure ordering of cluster hosts is consistent for same data
Python dictionaries don't have the order of their keys guaranteed, and
so when they are used to write config files, the order of the entries
may change. This can cause a restart of a dependent service.  Due to the
way that reactive invokes all 'true condition' handlers for every hook
invocation, this means that services might restart for every single hook
including 'update-status'.

This patch attempts to address that issue.

Related Bug: #1698814

Change-Id: Ic6c1d3d103a25b15ea202638c21213aea11cd3da