Add worker-multiplier for reactive charms
This adds the library code for calculating the number of workers that a
managed service should use.  It uses the
charmhelper.contrib.openstack.context helper classes to ensure that any
changes to those get reflected into reactive OpenStack charms.

The config option for worker-multiplier is contained in
laper-openstack-api's config.yaml.  Note that it is up to the charm
author to decide whether to use the simpler 'options.workers' or the
more complex 'options.wsgi_worker_context' which is a dictionary
returned from WSGIWorkerConfigContext().

Change-Id: I3cea350e536306655f5f109ec67ae7f0fba35fda