Rework the default handlers into layer-openstack-*
Due to commit 95493a4 in charms.reactive, the default handlers in
charms.openstack broke.  This is because charms.reactive no longer will
run handlers that are not in the hooks/ or reactive/ directory tree; the
default handlers in charms.openstack are in the library code, which is
pip installed into the module packages.

This patch, and related patches in layer-openstack and
layer-openstack-api enable the default handlers to function again.  Note
that from a charm author perspective, the API to default handlers is
identical.  This is merely an implementation change.

Change-Id: I7786a39ab29ab71a4cdb8bd652e8927cbd5fcf5c
Partial-Bug: #1707685