Fix Apache not being installed/configured for ssl
The configuration for SSL was not being run due to the vhost
not having been rendered and the ssl.enabled state not being
set when configure_ssl was run. This change creates a new state
ssl.requested which signals that ssl configuration needs to take
place. Also, an empty vhost is now created so a2ensite does not
fail.

Change-Id: Iaadcd30191e94f9345c55d9eb8551a5593b0fde4
Partial-Bug: #1719331