Add snap support
Update core charms.openstack classes to support installation
from snap as an alternative to installation from deb packages.

Introduce a new property 'package_type' supported on the
OpenStackCharm base class which can be set to 'deb' (default) or
'snap' to indicate which type of install  a specific OpenStackCharm
subclass supports.  The list of snaps is detailed in the 'snaps'
property, and the mode for installation of snaps is configured using
the 'snap_mode' property; as for packages, a specific snap can be
chosen to drive the application version that is used for the
deployed workload.

This is supported by an additional default package type selector
which, in conjunction with the existing release selector, will
pick the correct subclass of the OpenStackCharm class to use for
the configured installation source.

Snap installation is configured using the openstack-origin
charm configuration option - for example:

    openstack-origin=snap:ocata/stable

If no suitable class can be discovered for the configured
origin, the charm will throw an error as it does today for
missing openstack release support for deb based installations.

The linked gerrit review is the first charm implementation of
snap support based on these set of changes.

Required-By: I464025a2b72aba8c31a4a97ade39d2b2980c3a92
Change-Id: Iea33a939a6422da94d3c2c5b9a0748a47bfde11a