Always register the VIP with keystone catalog
Resolve_address is used to register an API charm's keystone catalog
entry. Resolve_address was checking if the unit was clustered and if not
falling back to its private-address. This is incorrect behavior. If a
VIP is set it should always register the VIP with the keystone catalog.

The fallout from this is manifested in two bugs: Bug#1750915 and
Bug#1749280. In first, keystone is creating and passing SSL
certificates for the local unit's private-ip and only subsequently for
the VIP cause apache to fail to start. In the second, a race condition
exists where clients of the API charm relying on catalog entries get
disrupted when the catalog entry changes.

This fix removes the is_clustered check from resolve_address. The
gating on is_clustered should happen elsewhere in the charm not at
registration time with keystone.

Change-Id: Ieaee4fbcd1b45b95ddca4ec3a9a09d4522b09b03
Partial-Bug: #1750915
Partial-Bug: #1749280