Ensure config files rendered with sane permissions
All configuration files are being rendered with root.root ownership
with mask 0444 (including global read!).  Render configuration files
as owner root, with a group read permission appropriate to the
consuming charm.  This can be configured by setting the group
attribute of the Charm subclass, but will default to 'root' if
not supplied (preserving the previous behaviour).

Change-Id: Ib1e2d3801b171e5c7ea79d058fb36dfc532d5d20
Closes-Bug: 1780490