Add interfaces from flags to adapter interface
When render_configs is called without an adapters_instance it used to
construct an adapters_instance which lacked any relation data. This
was, in effect, useless (the pre-existing docstring comment says as
much). Usually render_configs is called via render_with_interfaces
which inturn has gets its interface list excplicitly passed in.
However, this approach is not possible when needing to trigger a
config render from a generic layer as the generic layer does not know
which interfaces to pass.

This change adds code to dynamically build the interface list from
the flags which have been set already.

Change-Id: I1958226576c73d0ce0e9fd4c1423f86addadb454