Deal with obsolete packages during charm upgrade
For Rocky, we switch the payload execution environent from py2->py3
post initial charm release; ensure that any existing py2 based
deployments are automatically switched over to py3 during any
charm upgrade.

This is implemented in the default upgrade-charm handler code; if
packages are removed as part of the upgrade, all services associated
with the charm will be restarted.

Change-Id: Ib5f11d21dd26cb2918f5d5044ed56d688bba3b5d
Closes-Bug: 1803451