Consistent ordering of ADDRESS_TYPES when traversing dict
charms_openstack.adapters.ADDRESS_TYPES is traversed when filling
in network split information for haproxy configuration file generation.
The ADDRESS_TYPES constant is built from the keys of a simple python
dictionary, which provides no guaranteed ordering. This causes the
charm to render slight variations of the haproxy.cfg file and the
haproxy service is restarted as a result.

This is fixed by simply sorting the keys to ensure consistent ordering.
The list is reverse sorted so that the iteration occurs in the order
of public -> internal -> admin.

Change-Id: I89b2ee42b5827d0fe5bbb2ff7051e1bb5bd08c63
Closes-Bug: #1800980