Allow overriding of source configuration option key
The charms.openstack library contains lots of useful code for
reactive charms which is reusable for other charms.

Allow direct consumers of the OpenStackCharm class to override
the key used for package source selection.  This is usefull for
non-OpenStack charms deploying binaries from UCA (for example
Ceph)

Retain 'openstack-origin' as a default.

Also add and default to a `py3` target in tox for developer
friendliness.

Change-Id: I67d2e5d68bc4da87bcaba813f99d3f782f1d9907