Fix broken OpenStack pkg version check
The current OpenStack version check uses a package which does not
have semantic versioning and is not listed on the dict of package
version in charmhelpers. This change fixes this by using a package
which does have semantic versioning and is in the dict of versions.
If no package is installed which fit these requirements then
os_release is relied on to fallback to another method (checking
openstack-origin).

Change-Id: I256b2782e83e2619108726eeb27505d5f1854284
Closes-Bug: #1741628