Explicitly use pymysql sqlalchemy driver for Stein+
The sqlalchemy package has switched back to using mysqldb as the
default driver for the mysql dialect in Ubuntu Disco (Stein). For
Icehouse -> Rocky, OpenStack packages will successfully use the
default sqlalchemy driver [1] when only 'mysql' is specified in the
connection string. For Stein, the mysqldb driver must be overridden
with 'mysql+pymysql' in connection strings.

[1] Based on the OpenStack release, the sqlalchemy package defaults
to the following drivers:
  Stein+: mysqldb
  Mitaka-Rocky: pymysql
  Icehouse: mysqldb

Change-Id: I1ccc7080f54eda1c12b51a24a488666f793d038e