plugins/ceph: Pass key data instead of interface object
The helper functions for configuring keys will be used with both
RelationBase- and Endpoint-based interfaces.  They have different
mechanics to get to relation attributes (function calls vs.
properties).

The only interface attribute we use is the key data, so just pass
that instead of the interace object and let it be the callers job
to unpack the data as the interface they use require.

Change-Id: If1d645f4708e27b724f93cac0e14431137c885d7