Use reactive flags for pause/resume/series-upgrade
Currently when a unit is paused, resumed or series upgrade commands are
run there are calls out to charmhelpers which uses local kv store to set
state.

There is currently no way to use a reactive handler (when, when_not) to
determine these states.

This change adds the reactive handler flags while maintaining the
previous method.

Change-Id: I1b27ebeced1ee56f1133b8b34806571ad1b6c9fe