Add a custom status check called after all other checks
At present, use of the ``custom_assess_status_check`` method will
effectively disable the frameworks other automatic checks.

Add a new method to be called as the last check.  This is useful
when a charm author wants to both make use of the automatic checks
and add extra information when the charm is active and ready.

Change-Id: Iaf419c7f06e5a4dfffa04fcbd4dc1a755714851f
Closes-Bug: #1853141