Use charm codenames for select release when provided
A charm class can provide a dictionary that maps package
versions to OpenStack releases.  This dictionary is
consulted when determining if a OpenStack upgrade is available.

However, when determining release for selection of charm class
the map is not consulted.  This patch teaches the default select
release handler how to do so.

Pass on source configuration key as specified by charm class
to charm-helpers where appropriate.

Change-Id: If7f35e7d00533599c3800aa4a22c51861017969e
Closes-Bug: #1863774