test_utils: handle non-idempotent handlers on reload
A pattern across our charms is to run
``charms_openstack.charm.use_defaults()`` at the global level
on top of the module encapsulating the reactive handlers
(i.e. ``src/reactive/charm_handlers.py``).

An example of an non-idempotent handler is the
``charm.default-select-release`` default handler, it keeps
global state and throws a RuntimeError on subsequent calls.

The way TestRegisteredHooks reloads the module makes it
impossible to mock this out, and I think the only solution
would be to mock this out in the test helper library itself.

Closes-Bug: #1864088
Change-Id: I9ebbda895a7a748d8efa4e23eb0c3bdf660fa759