release select: Handle None return from get_os_codename_package
Change If7f35e7d00533599c3800aa4a22c51861017969e introduced a bug
that roots in the assumption that the ``get_os_codename_package``
function when called with ``fatal=True`` would raise an exception
in the event package or major version was not present in the map
provided to the function.

Instead of changing the API of ``get_os_codename_package`` method
we detect a ``None`` return in the default select handler.

Change-Id: Icad1811d1f183da25872b81192b69284373fbefb
Related-Bug: #1863774