Avoid multiple calls to create_ip_cert_links
Prior to this change, the call to create_ip_cert_links was inside the
for loop of tls objects. This caused LP Bug#1866741 where a sym link
with the same certificate name would get overwritten with a new cert.
This meant the last certificate of [internal, admin, public] would be
written to all three.

This change calls create_ip_cert_links and state sets only once. It also
adds a warning log entry if we accidentally overwrite sym linked files in
the future.

Change-Id: I4b36d7fd5e74ab3a894eb4a2451b5876fd777552
Partial-Bug: #1866741