Allow release selection to work for subordinate charms
The release selection function bases its decision on installed
packages and falls back to looking up OpenStack release codename
from source configuration option when no packages are installed.

Subordinate charms do not have a source configuration option,
thus we want to make it possible for a charm author to request
the release selection decision be made based on what is currently
available in the system APT cache alone.

Our principle charm should already have configured the system
with the desired UCA pocket enabled.

The charm author requests this behaviour by setting the
``source_config_key`` to an empty string.

Note that functions for upgrading between releases etc also
depend on the charm having a source configuration option but
that is OK as the subordinate has no business dealing with
that and should defer such operations to its principle charm.

Change-Id: Ieae81619e06dc287c2c42ec944425be4de13d0d6
Closes-Bug: #1865510