Allow specifying config key in call to configure_source
A charm may be in need of adding multiple sources (for
example vendor specific PPAs), and it would be useful
if the ``configure_source`` method would allow to do
this.

In such a use case overriding the ``source_config_key``
class variable would not be practical, so we allow
to optionally specify it as a argument to the method.

Change-Id: Ia5a8f05e863251520771a1c01a5f7dddf4be34d0
Closes-Bug: #1876720