Allow service actions to be overridden
In order to allow child classes to use other means of controlling
service lifetime, this change introduces methods that call
systemd/upstart via charm-helpers by default but allow for something
else to be used in overrides.

The example use-case in question is the use of crm tools to restart
services:

crm_resource --restart --wait --node <node-name>

Which will restart a resource on a given node if it is running there or
return ENXIO otherwise.

Change-Id: Ifb16f6743296a1ef6dcb212cad517afd57270f7f
Closes-Bug: #1891160