trilio: Support multiple NFS share configurations
TrilioVault supports use of multiple NFS shares using a comma
separated list.  Update the ghost-nfs-share action code to
support this type of configuration.

The assumption is that the order of shares provided to the
action is the same as the order of shares in the nfs-shares
configuration option in terms of replication between sites.

Mismatched length lists will generate an exception which can
be propagated back to the operator.

Change-Id: I7d501a3e44d4e7c732e7625fa2a8cde9a34ff564
Closes-Bug: 1894817