Optionally change service port for ssl
If a charm class sets ssl_port_bump to True then when configuring
the service for TLS also change the port it is listening on.

The need for this comes from the nova-scheduler and nova-compute
services which cache the url for the placement service, when the
placement service switches protocol the nova services fail which
does not trigger them to reexamine the catalogue.

Closes-Bug: #1862974

Change-Id: I790c55e78439b9a5ea2c3f12edbe5928a4f08e05