Further fix to OpenStackAPIRelationAdapters
The existing method of using a property to lazily load the cluster
relation adapter permanently changes the class to include that as
a cluster property.  This means that a future (in the same hook
execution) instantiation of the class would include that property,
which means that 'real' cluster adapter could not be used. Sadly,
this was happening with the manila charm, and possibly others.

This patch changes it from a lazily loaded property, to a simple
attribute.

Change-Id: I704f362e0dd845ff00d0f0470b0235a4dead5f9f
Closes-Bug: #1981736
(cherry picked from commit 581300632b426e985bf3daaed426daf3c89f825a)