Support wrapped release names
Charm instances where the release is > zed does not work using the
default_get_charm_instance. The keys stored in the _releases dictionary
are sorted in their alphabetical sorting order, thus causing releases
such as Bobcat to be treated as older releases than Zed.

This change causes the release names to be sorted using the
CompareOpenStackReleases object from the charmhelpers library. This will
ensure that the release names are sorted according the release order
rather than the alphabetical name.

Change-Id: I528a9dcc5511b74186132b57948e18e4f3a1ef0b