Support same pkg version for multiple releases
It is possible for the version of a package to remain the same
across multiple releases of openstack so this adds support to
get_os_codename_package() to allow this.

If releases A, B and C have version N and we are on release B
this will return release C so that it does not look like there
is an upgrade necessary. The main change here that allows this
is to support matching major and minor version up to X.Y.Z as
opposed to previously only matching the major version X.

Modifies one unittest to fix compatibility from unittest.mock
to mock.

Related-Bug: #1973303

Change-Id: I138f61312efb728544276483b1a459b9eeecafdb
(cherry picked from commit 55f98e80e430b1112c38b90d88e10bd0923daa6f)
(cherry picked from commit e58bad36743ecd1de41258b3c271f2c4c554a093)